Added comments and several procedures .
Updated the Ball class with comments and added an overload of IncrementSpeed to give a fixed amount.

Renamed Game1.cs to PungGame.cs.

Updated the Pallet class with comments and an overload of Increment speed to give a fixed amount instead.

Updated the Player class with comments and overloaded the AddScore to give it a fixed amount instead.

Replaced the Ball and Pallet bitmap images to .png to support transparency in the future.

Ball.png is not a square anymore but a circle. This was causing serious concerns among the users and was resolved. Move along.